requirements.txt
asgiref==3.7.2
Babel==2.14.0
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.10
django-cors-headers==4.1.0
django-environ==0.11.2
django-extra-views==0.14.0
django-haystack==3.2.1
django-oscar==3.2.4
django-phonenumber-field==6.4.0
django-rest-auth==0.9.5
django-tables2==2.3.4
django-treebeard==4.7.1
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.1.0
docopt==0.6.2
factory-boy==3.2.1
Faker==23.2.0
gunicorn==21.2.0
idna==3.6
packaging==23.2
phonenumbers==8.13.30
pillow==10.2.0
pipreqs==0.4.13
psycopg2-binary==2.9.5
purl==1.6
pycountry==23.12.11
PyJWT==2.8.0
pysolr==3.9.0
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
six==1.16.0
sorl-thumbnail==12.9.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.0
yarg==0.1.9
drf_yasg==1.21.7
